Fix issue where poDoctrineMigration overrides first import
When the first require is made (line 150), two classes are added to the declared list - the class defined in the migration file AND "poDoctrineMigration".

When the `end` function is called, sometimes we pick off the wrong newly generated class and assign the migration the name "poDoctrineMigration". 

This is not correct.